dsx: hardcode REG_ROMCTRL bits in dsxSendCommand
In the event that the driver is initialized without booting the
cartridge, reading the cartridge header from memory would cause
undefined behaviour, as the header may have different values set.

Also, the previous implementation would have broke in DSi mode, as it
was reading a no-longer-mirrored memory region. Which would have worked
just fine in DS mode... just not in DSi.

DS-Xtreme has been defunct for over a decade, so hardcoding it this way
is safe, as there are no more firmware updates.