scdssdhc2: major refactoring to support quirky carts
Normally, a command for this cart is like so:
`80 00 00 00 00 00 00 00`

The 80 command would ignore the remaining bytes, as it is simply here
to wait for the data buffer to be filled.

However, there are carts where if one were to send this command
beforehand to read a single block:
`53 DE AD BE EF 00 00 00`

and then use the 80 command to wait for data, it sends this instead:
`80 DE AD BE EF 00 00 00`

The original SuperCard code does in fact do this, as a way of reducing
the size of the binary? All other DSTT based carts do not care about
the remaining bytes, but some clone manufacturers appear to have taken
the commands *too* literally.

libtwl wasn't really designed for this use case, so instead define our
own card command register casted to a u8 array, and reorganize commands
to work in a similar manner to the original SuperCard design.